+++ src/pages/feed-order-page.module.css
@@ -1,0 +1,11 @@
+.content {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  width: 640px;
+  align-items: center;
+}
